requirements.txt
# requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
pydantic-settings==2.2.1
transformers==4.40.0
torch==2.6.0
# googletrans==4.0.0-rc1
google-generativeai==0.5.2
python-dotenv==1.0.1
redis==5.0.1
sqlalchemy==2.0.29
celery==5.3.4
# httpx==0.27.0
sentencepiece==0.1.99

# ML/Data Science packages for analytics
pandas==2.2.0
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.8.0
seaborn==0.12.2
joblib==1.3.2

# Additional analytics packages
plotly==5.17.0
scipy==1.11.3